+++ postgres_metrics/models.py
@@ -1,0 +1,13 @@
+from django.db import models
+
+from .metrics import registry
+
+
+class Metric(models.Model):
+
+    class Meta:
+        default_permissions = []
+        permissions = [
+            (metric.permission_name, 'Can view metric %s' % metric.label)
+            for metric in registry
+        ]